Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -49,7 +49,7 @@ int main()
     {
         if (b != 0)  // To prevent division by zero
         {
-            printf("%d", (float)a / b);  // Division (output as float)
+            printf("%d", a / b);  // Division (output as float)
         }
         else
         {
